use shared BlockingView
